Move order-row to separate module
