+++ src/app.module.ts
@@ -11,9 +11,9 @@ import { ProductsModule } from "./products/products.module";
 
 import { User } from "./users/user.entity";
 import { Product } from "./products/product.entity";
-import { Order } from "./orders/entities/order.entity";
+import { Order } from "./orders/order.entity";
 import { OrdersModule } from "./orders/orders.module";
-import { OrderRow } from "./orders/entities/order-row.entity";
+import { OrderRow } from "./order-rows/order-row.entity";
 import { Category } from "./categories/category.entity";
 
 @Module({
