+++ src/orders/enum/order-status.enum.ts
@@ -1,4 +1,4 @@
-export enum StatusEnum {
+export enum OrderStatusEnum {
   CREATED = 'created',
   PLACED = 'placed',
   IN_PROGRESS = 'in progress',
