+++ src/orders/orders.module.ts
@@ -2,18 +2,21 @@ import { Module } from "@nestjs/common";
 
 import { TypeOrmModule } from "@nestjs/typeorm";
 
-import { Order } from "./entities/order.entity";
+import { Order } from "./order.entity";
 import { OrdersService } from "./orders.service";
 import { OrdersController } from "./orders.controller";
-import { OrderRow } from "./entities/order-row.entity";
 import { Product } from "../products/product.entity";
+import { OrderRowModule } from "../order-rows/order-row.module";
 
 @Module({
   providers: [
     OrdersService,
   ],
   controllers: [OrdersController],
-  imports: [TypeOrmModule.forFeature([Order, OrderRow, Product])],
+  imports: [
+    OrderRowModule,
+    TypeOrmModule.forFeature([Order, Product])
+  ],
   exports: [OrdersService],
 })
 export class OrdersModule {
